+++ src/eggbot/core_entities.py
@@ -8,7 +8,9 @@ import json
 import os
 import logging
 import pathlib
-import typing
+
+from typing import Any
+from typing import Callable
 
 import dotenv
 
@@ -108,11 +110,11 @@ class CoreConfig(object):
             return False
         return True
 
-    def read(self, key: str) -> typing.Any:
+    def read(self, key: str) -> Any:
         """ Reads values by key from config. Returns None if not exists """
         return self.__config.get(key)
 
-    def create(self, key: str, value: typing.Any = None) -> bool:
+    def create(self, key: str, value: Any = None) -> bool:
         """ Creates a key/value pair in the configuration """
         if not isinstance(key, str):
             logger.error(f".create() key not string. Given a {type(key)}")
@@ -123,7 +125,7 @@ class CoreConfig(object):
         self.__config[key] = value
         return True
 
-    def update(self, key: str, value: typing.Any = None) -> bool:
+    def update(self, key: str, value: Any = None) -> bool:
         """ Updates key/value pair in the configuration """
         if key not in self.__config.keys():
             logger.error(".update() key not found, use .create()")
@@ -137,4 +139,113 @@ class CoreConfig(object):
             logger.error(".delete() key not found.")
             return False
         del self.__config[key]
+        return True
+
+
+class EventSub(object):
+    """Allows modules to sub to event notifications
+
+    A component to the pub/sub framework for how the bot handles events
+    and allows future scalability. Declare one EventSub() object per
+    bot instance. Multiple declared instances can be used but do not
+    provide additional control or function.
+    """
+
+    def __init__(self):
+        self._pubsub = {}
+
+    def sub_create(self, target: Callable, event: str) -> bool:
+        """Subscribe a target to an event
+
+        Args:
+            target - The target callable object
+            event - The name of the event group to be subbed
+
+        Returns:
+            Bool - False if failed to create (check logs)
+        """
+        if not isinstance(target, Callable):
+            logger.error(f".sub(), [target] is not callable: {type(target)}")
+            return False
+
+        if not self._pubsub.get(event):
+            self._pubsub[event] = set()  # Ensure we have a key
+
+        if target in self._pubsub[event]:
+            logger.error(".sub(), [target] already subscribed")
+            return False
+
+        self._pubsub[event].add(target)
+
+        return True
+
+    def sub_delete(self, target: callable, event: str) -> bool:
+        """Unsubscribe a target from an event
+
+        Args:
+            target - The target callable object
+            event - The name of the event group to be unsubbed
+
+        Returns:
+            Bool - False if failed to unsub (check logs)
+        """
+        if not self._pubsub.get(event):
+            logger.error(f".unsub(), Event '{event}' does not exist.")
+            return False
+
+        if target not in self._pubsub[event]:
+            logger.error(f".unsub(), [target] is not subbed to '{event}'")
+            return False
+
+        self._pubsub[event].remove(target)
+
+        return True
+
+    def sub_wipe(self, target: callable) -> bool:
+        """Removes target from all subscriptions
+
+        Args:
+            target - The target callable object to be removed
+
+        Returns:
+            True
+        """
+        events = self.event_list_all()
+
+        for event in events:
+            self._pubsub[event].discard(target)
+
+        return True
+
+    def sub_list(self, target: callable) -> tuple:
+        """ Returns a list of events that target is subbed to """
+        sub_list = []
+        for event in self.event_list_all():
+            if target in self._pubsub[event]:
+                sub_list.append(event)
+        return tuple(sub_list)
+
+    def event_list_all(self) -> tuple:
+        """ Returns tuple of all events """
+        return tuple(self._pubsub.keys())
+
+    def event_list(self, event: str) -> tuple:
+        """ Returns all subscribed callables in event """
+        if event not in self._pubsub.keys():
+            logger.warning(f'event_list(), Event "{event}" not found.')
+        return tuple(self._pubsub.get(event, ()))
+
+    def event_delete(self, event: str) -> bool:
+        """ Deletes an entire event, unsubbing all within """
+        if event not in self._pubsub.keys():
+            logger.error(f'event_delete(), Event "{event}" not found.')
+            return False
+        del self._pubsub[event]
+        return True
+
+    def event_create(self, event: str) -> bool:
+        """ Creates an empty event set """
+        if event in self._pubsub.keys():
+            return False
+        self._pubsub[event] = set()
         return True